+++ src/pages/home/accounts.js
@@ -1,0 +1,44 @@
+import { IdService } from '../../services';
+
+// @ts-nocheck
+class AccountsPage extends HTMLElement {
+    constructor() {
+        super();
+        this.shadow = this.attachShadow({mode: 'closed'});
+        this.active = this.getAttribute('active') || 'false';
+    }
+    
+    static get observedAttributes() { 
+        return ['active']; 
+    }
+
+    connectedCallback() {
+        this.render();
+    }
+
+    attributeChangedCallback(name, oldValue, newValue) {
+        const el = IdService.id('base-home', this.shadow);
+        el?.setAttribute('active', newValue);
+    }
+
+    render() {
+        this.shadow.innerHTML = `
+            <style>
+                .account {
+                    background-color: white;
+                }
+                .banner {
+                    padding-bottom: 10px;
+                }
+            </style>
+            <home-account></home-account>
+            <section class="banner">
+                <home-banners></home-banners>
+            </section>
+        `;
+    }
+}
+
+if ('customElements' in window) {
+	customElements.define('home-accounts-page', AccountsPage);
+}