+++ src/pages/home/index.js
@@ -1,3 +1,4 @@
 import './account';
 import './banners';
 import './homeBanner';
+import './accounts';