+++ src/pages/start/sets.js
@@ -1,0 +1,7 @@
+export const HomePageTabs = {
+    tabLinks: {
+        accounts: 'account-btn',
+        welcome: 'welcome-btn',
+        start: 'start-btn',
+    }
+};